change variable type

+++ src/main/java/Fruitables/shop/dto/ProductDetailDTO.java
@@ -19,11 +19,11 @@ public class ProductDetailDTO {
     private CategoryDTO category;
     private String name;
     private int qtyInStock;
-    private BigDecimal price;
+    private double price;
     private String description;
     private List<String> image;
 
-    public ProductDetailDTO(CategoryDTO category, String name, int qtyInStock, BigDecimal price, String description, List<String> image) {
+    public ProductDetailDTO(CategoryDTO category, String name, int qtyInStock, double price, String description, List<String> image) {
         this.category = category;
         this.name = name;
         this.qtyInStock = qtyInStock;
